# backend/docs/openapi_endpoints.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "FERI-AD Microservices Endpoints",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "https://mkc-backend-kqov.onrender.com", "description": "Node API (production)" },
    { "url": "https://mkc-5slv.onrender.com/api/v1", "description": "Python service (production)" }
  ],
  "paths": {
...
    "/users": { "get": { "summary": "List users (admin)", "security": [{ "BearerAuth": [] }] } },
...
    "/requests": { "post": { "summary": "Create request (client)", "security": [{ "BearerAuth": [] }] } },
...
    "/documents": { "get": { "summary": "List all documents (admin)", "security": [{ "BearerAuth": [] }] } },
    "/admin/users": { "get": { "summary": "List users (admin)", "security": [{ "BearerAuth": [] }] } },
...
    "/admin/requests": { "get": { "summary": "List requests (admin)", "security": [{ "BearerAuth": [] }] } },
...
    "/admin/documents": { "get": { "summary": "List documents (admin)", "security": [{ "BearerAuth": [] }] } },
...
    "/notifications": { "get": { "summary": "List notifications (auth)", "security": [{ "BearerAuth": [] }] } },
...
    "/api/v1/parse/document": { "post": { "summary": "Parse document (API key required)", "security": [{ "ApiKeyAuth": [] }] } },
    "/api/v1/generate/feri": { "post": { "summary": "Generate FERI PDF (API key required)", "security": [{ "ApiKeyAuth": [] }] } },
    "/api/v1/generate/ad": { "post": { "summary": "Generate AD PDF (API key required)", "security": [{ "ApiKeyAuth": [] }] } }
  },
  "components": {
    "securitySchemes": {
      "BearerAuth": { "type": "http", "scheme": "bearer" },
      "ApiKeyAuth": { "type": "apiKey", "in": "header", "name": "x-api-key" }
    }
  }
}